\name{getK}
\docType{methods}
\alias{getK}
\alias{getK,CDEPD-method}

\title{Method for retrieving the count of scenarios}

\description{
  Additional arguments to the \code{getK}-method pertinent to the
  defined S4-classes in this package are detailed below.
}

\usage{
\S4method{getK}{CDEPD}(object)  
}

\arguments{
  \item{object}{\code{CDEPD}: ordinarily returned by a call to \code{DEP()}}
}

\value{\code{integer}: Returns the count of scenarios.
}

\author{Bernhard Pfaff}

\keyword{methods}
