\name{getTech}
\docType{methods}
\alias{getTech}
\alias{getTech,CDEPD-method}

\title{Method for retrieving list of technology matrices}

\description{
  Additional arguments to the \code{getTech}-method pertinent to the
  defined S4-classes in this package are detailed below.
}

\usage{
\S4method{getTech}{CDEPD}(object)  
}

\arguments{
  \item{object}{\code{CDEPD}: ordinarily returned by a call to \code{DEP()}}
}

\value{\code{list}: Elements are the technology matrices employed in the
  scenarios. 
}

\author{Bernhard Pfaff}

\keyword{methods}
